+++ R/RcppExports.R
@@ -32,6 +32,18 @@ rgb_2gray <- function(RGB_image) {
     .Call(`_OpenImageR_rgb_2gray`, RGB_image)
 }
 
+Gabor_Filter_Bank <- function(u, v, m, n, plot_data = FALSE) {
+    .Call(`_OpenImageR_Gabor_Filter_Bank`, u, v, m, n, plot_data)
+}
+
+Gabor_export_Features <- function(img, d1, d2, u, v, m, n, downsample_vec = FALSE, plot_data = FALSE, normalize_features = FALSE, threads = 1L) {
+    .Call(`_OpenImageR_Gabor_export_Features`, img, d1, d2, u, v, m, n, downsample_vec, plot_data, normalize_features, threads)
+}
+
+Gabor_generate <- function(img_data, img_nrow, img_ncol, d1, d2, u, v, m, n, downsample_vec = FALSE, normalize_features = FALSE, threads = 1L) {
+    .Call(`_OpenImageR_Gabor_generate`, img_data, img_nrow, img_ncol, d1, d2, u, v, m, n, downsample_vec, normalize_features, threads)
+}
+
 hog_cpp <- function(image, n_divs = 3L, n_bins = 6L) {
     .Call(`_OpenImageR_hog_cpp`, image, n_divs, n_bins)
 }
@@ -206,6 +218,10 @@ rotate_rcpp <- function(image, angle) {
 
 seq_rcpp_range <- function(start, end) {
     .Call(`_OpenImageR_seq_rcpp_range`, start, end)
+}
+
+inner_bilinear <- function(t, s, image, method, n, m, mm, nn, thet) {
+    .Call(`_OpenImageR_inner_bilinear`, t, s, image, method, n, m, mm, nn, thet)
 }
 
 rotate_nearest_bilinear <- function(image, angle, method, mode, threads) {